+++ app/src/main/res/layout/fragment_diary_calendar.xml
@@ -166,11 +166,10 @@
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_calendar"
             android:layout_width="match_parent"
-            android:layout_height="0dp"
+            android:layout_height="wrap_content"
             android:paddingHorizontal="16dp"
             android:background="@color/light_gray_1"
-            app:layout_constraintTop_toBottomOf="@id/lo_calendar_weekdays"
-            app:layout_constraintBottom_toTopOf="@id/lo_bottom" />
+            app:layout_constraintTop_toBottomOf="@id/lo_calendar_weekdays" />
 
         <ImageView
             android:layout_width="match_parent"
